incompleted revision

+++ O2DESNet/Optimizers/Optimizer.cs
@@ -11,22 +11,22 @@ namespace O2DESNet.Optimizers
         where TSimulator : Simulator<TScenario, TStatus>
     {
         protected DecisionSpace DecisionSpace { get; private set; }
-        protected Func<double[], TScenario> ConstrScenario { get; private set; }
+        protected Func<decimal[], TScenario> ConstrScenario { get; private set; }
         protected bool Discrete { get; private set; }
         protected Random DefaultRS { get; private set; }
-        internal Dictionary<ArrayKey<double>, TScenario> Scenarios { get; private set; } // lookup for scenario given decision
-        internal Dictionary<TScenario, double[]> Decisions { get; private set; } // lookup for decision given scenario
+        internal Dictionary<ArrayKey<decimal>, TScenario> Scenarios { get; private set; } // lookup for scenario given decision
+        internal Dictionary<TScenario, decimal[]> Decisions { get; private set; } // lookup for decision given scenario
         public MinSelector<TScenario, TStatus, TSimulator> Replicator { get; private set; }
         public TScenario Optimum { get { return Replicator.Optimum; } }
 
         public Optimizer(
             DecisionSpace decisionSpace,
-            Func<double[], TScenario> constrScenario,
+            Func<decimal[], TScenario> constrScenario,
             Func<TScenario, int, TStatus> constrStatus,
             Func<TStatus, TSimulator> constrSimulator,
             Func<TStatus, bool> terminate,
-            Func<TStatus, double> objective,
-            double inDifferentZone = 0,
+            Func<TStatus, decimal> objective,
+            decimal inDifferentZone = 0,
             bool discrete = false,
             int seed = 0)
         {
@@ -34,8 +34,8 @@ namespace O2DESNet.Optimizers
             ConstrScenario = constrScenario;
             Discrete = discrete;
             DefaultRS = new Random(seed);
-            Scenarios = new Dictionary<ArrayKey<double>, TScenario>();
-            Decisions = new Dictionary<TScenario, double[]>();
+            Scenarios = new Dictionary<ArrayKey<decimal>, TScenario>();
+            Decisions = new Dictionary<TScenario, decimal[]>();
             Replicator = new MinSelector<TScenario, TStatus, TSimulator>(
                 new TScenario[] { }, constrStatus, constrSimulator, terminate, objective, inDifferentZone);
         }
@@ -49,7 +49,7 @@ namespace O2DESNet.Optimizers
             {
                 var decision = decisions[i];
                 if (Discrete) decision = decision.Select(d => Math.Round(d)).ToArray(); // discretize
-                var key = new ArrayKey<double>(decision);
+                var key = new ArrayKey<decimal>(decision);
                 if (!Scenarios.ContainsKey(key))
                 {
                     // create and include new scenario
@@ -61,7 +61,7 @@ namespace O2DESNet.Optimizers
             }
             Alloc(budget - countNewScenarios * Replicator.InitBudget);
         }
-        protected virtual List<double[]> Sample(int size) { return DecisionSpace.Sample(size, DefaultRS); }
+        protected virtual List<decimal[]> Sample(int size) { return DecisionSpace.Sample(size, DefaultRS); }
         protected virtual void Alloc(int budget) { Replicator.OCBAlloc(budget); }
     }       
 }
