+++ O2DESNet/Simulator.cs
@@ -76,8 +76,8 @@ namespace O2DESNet
         #region For Time Dilation
         private DateTime _realTimeAtDilationReset;
         private DateTime _dilatedTimeAtDilationScaleReset;
-        private double _timeDilattionScale;
-        public double TimeDilationScale
+        private decimal _timeDilattionScale;
+        public decimal TimeDilationScale
         {
             get { return _timeDilattionScale; }
             set
